% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{A}
\alias{A}
\title{Example SNP-to-cell annotation matrix}
\format{
A numeric matrix with 200 rows and 8 columns.
}
\source{
Simulated data.
}
\usage{
A
}
\description{
Simulated annotation matrix linking SNPs to cell types or functional annotations.
}
\keyword{datasets}
