% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{B}
\alias{B}
\title{Example gene-to-cell annotation matrix}
\format{
A numeric matrix with 60 rows and 8 columns.
}
\source{
Simulated data.
}
\usage{
B
}
\description{
Simulated annotation matrix linking genes to cell types or functional annotations.
}
\keyword{datasets}
