% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesMAP-package.R
\docType{package}
\name{BayesMAP-package}
\alias{BayesMAP}
\alias{BayesMAP-package}
\title{BayesMAP}
\description{
Bayesian multi-level association mapping package.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

Authors:
\itemize{
  \item First Last \email{first.last@example.com}
}

}
\keyword{internal}
