% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{L}
\alias{L}
\title{Example SNP-to-gene annotation matrix}
\format{
A numeric matrix with 200 rows and 60 columns.
}
\source{
Simulated data.
}
\usage{
L
}
\description{
Simulated annotation matrix linking SNPs to genes.
}
\keyword{datasets}
