% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{X}
\alias{X}
\title{Example genotype matrix}
\format{
A numeric matrix with 300 rows and 200 columns.
}
\source{
Simulated data.
}
\usage{
X
}
\description{
Simulated genotype or design matrix used in BayesMAP examples.
}
\keyword{datasets}
