% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{y}
\alias{y}
\title{Example phenotype vector}
\format{
A numeric vector of length 300.
}
\source{
Simulated data.
}
\usage{
y
}
\description{
Simulated quantitative phenotype used in BayesMAP examples.
}
\keyword{datasets}
