% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesMAP-package.R
\docType{package}
\name{BayesMAP-package}
\alias{BayesMAP}
\alias{BayesMAP-package}
\title{BayesMAP}
\description{
Bayesian multi-level association mapping package.
}
\author{
\strong{Maintainer}: Moksedul Momin \email{cvasu.momin@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Moksedul Momin \email{cvasu.momin@gmail.com} [copyright holder]
  \item Jian Zeng \email{j.zeng@imb.uq.edu.qu} [copyright holder]
}

}
\keyword{internal}
